Address review feedback to improve eval server behavior
